+++ src/Kiota.Builder/Refiners/PhpReservedNamesProvider.cs
@@ -68,7 +68,8 @@ namespace Kiota.Builder.Refiners
             "xor",
             "yield",
             "yield from",
-            "BaseRequestBuilder"
+            "BaseRequestBuilder",
+            "MultipartBody"
         });
 
         public HashSet<string> ReservedNames => _reservedNames.Value;